bench: add binary to compute serialized size
